Split out console.c and sc6551.c from acia.c
getting ready to add tcp stuff but need to fix some things first
split to make the code more manageable
Added some how to use text to ReadMe
