+++ acia/ReadMe.txt
@@ -2,47 +2,54 @@
     DYNAMIC LINK LIBRARY : acia Project Overview
 ========================================================================
 
-AppWizard has created this acia DLL for you.
-
-This file contains a summary of what you will find in each of the files that
-make up your acia application.
 
 
-acia.vcxproj
-    This is the main project file for VC++ projects generated using an Application Wizard.
-    It contains information about the version of Visual C++ that generated the file, and
-    information about the platforms, configurations, and project features selected with the
-    Application Wizard.
+Using /t2 on (Nitr)Os9
 
-acia.vcxproj.filters
-    This is the filters file for VC++ projects generated using an Application Wizard. 
-    It contains information about the association between the files in your project 
-    and the filters. This association is used in the IDE to show grouping of files with
-    similar extensions under a specific node (for e.g. ".cpp" files are associated with the
-    "Source Files" filter).
+Terminal settings for Windows console should be as follows:
 
-acia.cpp
-    This is the main DLL source file.
+/t2
+ -upc -bsb bsl echo lf null=0 -pause pag=32 bsp=08
+del=18 eor=0D eof=1B reprint=09 dup=19 psc=17 abort=03
+quit=05 bse=08 bell=07 type=80 baud=06 xon=11 xoff=13
 
-	When created, this DLL does not export any symbols. As a result, it
-	will not produce a .lib file when it is built. If you wish this project
-	to be a project dependency of some other project, you will either need to
-	add code to export some symbols from the DLL so that an export library
-	will be produced, or you can set the Ignore Input Library property to Yes
-	on the General propert page of the Linker folder in the project's Property
-	Pages dialog box.
+Basically the idea is to match CoCo window settings.  Acia.dll
+translates the control codes to the proper console functions.
 
-/////////////////////////////////////////////////////////////////////////////
-Other standard files:
+Colors default to white on black.  Only colors 0-7 are supported.
 
-StdAfx.h, StdAfx.cpp
-    These files are used to build a precompiled header (PCH) file
-    named acia.pch and a precompiled types file named StdAfx.obj.
+The console handles scrolling fairly well. The windows system 
+automatically moves the screen within the screen buffer to keep
+the cursor visible. This causes text to appear to scroll as long
+as the cursor is within the boundaries of the buffer. The vertical
+scroll bar allows the user to scroll back into the buffer to see
+what was printed previously.
 
-/////////////////////////////////////////////////////////////////////////////
-Other notes:
+By default the windows scroll buffer is large but eventually it 
+becomes full. To handle this acia.dll deletes lines from the top of
+the buffer when the cursor tries to move to a line below the bottom.
+This affects full screen editors.  They use cursor move sequences
+to control positioning within the screen boundaries but must scroll
+text when the cursor tries to move beyond the screen boundaries.
 
-AppWizard uses "TODO:" comments to indicate parts of the source code you
-should add to or customize.
+Vim (fixed tsedit) works pretty well. It takes care of deleting lines
+from the top of the screen when the cursor nears the bottom and
+replaces the lines when scrolling up. Vim seems to be hardcoded to
+24 lines so the screen must be at least 24 lines long. (changing
+tspars or typing v to set the screen type does not work for me)
 
-/////////////////////////////////////////////////////////////////////////////
+Scred assumes the terminal will delete lines from the top of the 
+screen when the cursor is at the bottom of the screen. (emphasis on
+screen) This means screen length must match what scred thinks for
+it to work properly.  Since acia.dll console does the delete based
+on buffer size it actually is that size that scred must be aware of.
+Fix it to set the screen buffer to a reasonable size and use scred's
+ -l=<screen buffer len> argument. (500 lines still seems to work 
+but I don't use scred much). The size of the buffer can be changed
+from the console properties menu.
+
+I did a brief check of uemacs. It seems to work okay too except I
+find the lack of support of arrow keys too annoying and I am too 
+used to vi's idea use of hjkl for cursor position to adapt to emacs.
+
+
