+++ acia/resource.h
@@ -1,17 +1,19 @@
 //{{NO_DEPENDENCIES}}
-// Microsoft Developer Studio generated include file.
+// Microsoft Visual C++ generated include file.
+// Used by acia.rc
 //
 #define IDS_MODULE_NAME                 1
 #define IDS_VERSION                     2
 #define IDS_CATNUMBER                   3
 #define IDD_PROPPAGE                    102
 #define IDC_TCPPORT                     1004
+#define IDC_TCPHOST                     1005
 
 // Next default values for new objects
 // 
 #ifdef APSTUDIO_INVOKED
 #ifndef APSTUDIO_READONLY_SYMBOLS
-#define _APS_NEXT_RESOURCE_VALUE        101
+#define _APS_NEXT_RESOURCE_VALUE        102
 #define _APS_NEXT_COMMAND_VALUE         40001
 #define _APS_NEXT_CONTROL_VALUE         1008
 #define _APS_NEXT_SYMED_VALUE           101
